% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alvin_mask.R
\name{alvin_mask}
\alias{alvin_mask}
\alias{alvin_image}
\title{ALVIN Mask}
\usage{
alvin_mask(mm = c("2", "1", "0.5"))

alvin_image(mm = c("2", "1", "0.5"), ...)
}
\arguments{
\item{mm}{The resolution for the MNI template to correspond to.}

\item{...}{additional arguments to pass to \code{\link{readnii}}}
}
\value{
Path to ALVIN NIfTI image
}
\description{
ALVIN Mask
}
\examples{
alvin_mask()
alvin_mask(mm = 1)
alvin_mask(mm = 0.5)
alvin_image()
}
